+++ lib/pylon.dart
@@ -5,4 +5,6 @@ export 'src/future.dart';
 export 'src/mutable_pylon.dart';
 export 'src/pylon.dart';
 export 'src/stream.dart';
+export 'src/port.dart';
+export 'src/codec.dart';
 export 'src/util/extensions.dart';